# Faster string encoding