--- a/json/encode.go
+++ b/json/encode.go
@@ -536,6 +536,8 @@ func (e encoder) encodeStruct(b []byte, p unsafe.Pointer, st *structType) ([]byt
 	var n int
 	b = append(b, '{')
 
+	escapeHTML := (e.flags & EscapeHTML) != 0
+
 	for i := range st.fields {
 		f := &st.fields[i]
 		v := unsafe.Pointer(uintptr(p) + f.offset)
@@ -544,7 +546,7 @@ func (e encoder) encodeStruct(b []byte, p unsafe.Pointer, st *structType) ([]byt
 			continue
 		}
 
-		if (e.flags & EscapeHTML) != 0 {
+		if escapeHTML {
 			k = f.html
 		} else {
 			k = f.json
@@ -553,11 +555,10 @@ func (e encoder) encodeStruct(b []byte, p unsafe.Pointer, st *structType) ([]byt
 		lengthBeforeKey := len(b)
 
 		if n != 0 {
-			b = append(b, ',')
+			b = append(b, k...)
+		} else {
+			b = append(b, k[1:]...)
 		}
-
-		b = append(b, k...)
-		b = append(b, ':')
 
 		if b, err = f.codec.encode(e, b, v); err != nil {
 			if err == (rollback{}) {