--- a/json/string.go
+++ b/json/string.go
@@ -1,6 +1,7 @@
 package json
 
 import (
+	"math/bits"
 	"reflect"
 	"unsafe"
 )
@@ -10,10 +11,12 @@ const (
 	msb = 0x8080808080808080
 )
 
-// simpleString checks whether `s` is made up of chars in the range [0x20, 0x7F]
-// and doesn't include a double quote or backslash. If the escapeHTML mode is
-// enabled, `s` cannot contain <, > or & either.
-func simpleString(s string, escapeHTML bool) bool {
+// escapeIndex finds the index of the first char in `s` that requires escaping.
+// A char requires escaping if it's outside of the range of [0x20, 0x7F] or if
+// it includes a double quote or backslash. If the escapeHTML mode is enabled,
+// the chars <, > and & also require escaping. If no chars in `s` require
+// escaping, the return value is -1.
+func escapeIndex(s string, escapeHTML bool) int {
 	chunks := stringToUint64(s)
 	for _, n := range chunks {
 		// combine masks before checking for the MSB of each byte. We include
@@ -24,37 +27,39 @@ func simpleString(s string, escapeHTML bool) bool {
 			mask |= contains(n, '<') | contains(n, '>') | contains(n, '&')
 		}
 		if (mask & msb) != 0 {
-			return false
+			return (bits.TrailingZeros64(mask&msb) + 1) / 8
 		}
 	}
 
 	for i := len(chunks) * 8; i < len(s); i++ {
 		c := s[i]
 		if c < 0x20 || c > 0x7f || c == '"' || c == '\\' || (escapeHTML && (c == '<' || c == '>' || c == '&')) {
-			return false
+			return i
 		}
 	}
 
-	return true
-}
-
-// expand puts the specified byte into each of the 8 bytes of a uint64.
-func expand(b byte) uint64 {
-	return lsb * uint64(b)
+	return -1
 }
 
 // below return a mask that can be used to determine if any of the bytes
 // in `n` are below `b`. If a byte's MSB is set in the mask then that byte was
-// below `b`. The result is only valid if `b` is below 0x80.
+// below `b`. The result is only valid if `b`, and each byte in `n`, is below
+// 0x80.
 func below(n uint64, b byte) uint64 {
 	return n - expand(b)
 }
 
 // contains returns a mask that can be used to determine if any of the
 // bytes in `n` are equal to `b`. If a byte's MSB is set in the mask then
-// that byte is equal to `b`. The result is only valid if `b` is below 0x80.
+// that byte is equal to `b`. The result is only valid if `b`, and each
+// byte in `n`, is below 0x80.
 func contains(n uint64, b byte) uint64 {
 	return (n ^ expand(b)) - lsb
+}
+
+// expand puts the specified byte into each of the 8 bytes of a uint64.
+func expand(b byte) uint64 {
+	return lsb * uint64(b)
 }
 
 func stringToUint64(s string) []uint64 {