This PR adds a fast path for strings that don't require escaping.

```
name           old time/op    new time/op    delta
CodeEncoder-4    2.33ms ± 2%    2.12ms ± 0%   -9.05%  (p=0.000 n=7+8)

name           old speed      new speed      delta
CodeEncoder-4   834MB/s ± 2%   917MB/s ± 0%   +9.93%  (p=0.000 n=7+8)
```

`encodeString` does a pass over the string to determine if it can copy it straight into the output buffer. A [SWAR](https://en.wikipedia.org/wiki/SWAR) technique derived from https://graphics.stanford.edu/~seander/bithacks.html#HasLessInWord is used to check 8 bytes at a time:

```
SimpleString4KB-4            3.91GB/s ± 0%
SimpleString4KBEscapeHTML-4  3.07GB/s ± 0%
SimpleString1-4               217MB/s ± 0%
SimpleString1EscapeHTML-4     195MB/s ± 0%
SimpleString7-4               683MB/s ± 0%
SimpleString7EscapeHTML-4     510MB/s ± 0%
```

I get 3-4GB/s which is not bad for a pass free of SIMD instructions.

I threw in an improvement to struct encoding too (https://github.com/segmentio/encoding/pull/50/commits/933a4e3ab3eec8b63166f0d7db8dcc6954553bf5) which accounts for 1-2% of the gains.