@chriso 2020-09-19
Current standings:

```
name                 old speed     new speed      delta
Marshal/string#01-4  397MB/s ± 0%   477MB/s ± 0%   +20.29%  (p=0.008 n=5+5)
Marshal/string#02-4  392MB/s ± 1%   503MB/s ± 1%   +28.20%  (p=0.002 n=6+6)
Marshal/string#03-4  399MB/s ± 1%   507MB/s ± 0%   +27.11%  (p=0.008 n=5+5)
Marshal/string#04-4  403MB/s ± 0%   506MB/s ± 1%   +25.57%  (p=0.002 n=6+6)
Marshal/string#05-4  404MB/s ± 2%   508MB/s ± 1%   +25.89%  (p=0.002 n=6+6)
Marshal/string#06-4  392MB/s ± 3%   486MB/s ± 0%   +24.12%  (p=0.002 n=6+6)
Marshal/string#07-4  497MB/s ± 1%   508MB/s ± 1%    +2.14%  (p=0.004 n=6+5)
Marshal/string#08-4  498MB/s ± 0%   505MB/s ± 1%    +1.57%  (p=0.004 n=5+6)
Marshal/string#09-4  222MB/s ± 7%   227MB/s ± 1%      ~     (p=1.000 n=6+6)
Marshal/string#10-4  134MB/s ± 0%   134MB/s ± 0%      ~     (p=0.105 n=4+6)
Marshal/string#11-4  134MB/s ± 1%   134MB/s ± 0%      ~     (p=0.132 n=6+6)
Marshal/string#12-4  134MB/s ± 0%   134MB/s ± 0%      ~     (p=0.952 n=5+5)
Marshal/string#13-4  355MB/s ± 1%   352MB/s ± 1%      ~     (p=0.132 n=6+6)
Marshal/string#14-4  205MB/s ± 6%   208MB/s ± 0%      ~     (p=0.931 n=6+5)
Marshal/string#15-4  208MB/s ± 1%   208MB/s ± 0%      ~     (p=0.690 n=5+5)
Marshal/string#16-4  358MB/s ± 1%   354MB/s ± 0%    -1.11%  (p=0.004 n=6+5)
Marshal/string#17-4  477MB/s ± 1%   370MB/s ± 0%   -22.50%  (p=0.004 n=6+5)
Marshal/string#18-4  589MB/s ± 0%   495MB/s ± 0%   -15.90%  (p=0.008 n=5+5)
Marshal/string#19-4  605MB/s ± 0%  1055MB/s ± 1%   +74.25%  (p=0.004 n=5+6)
Marshal/string#20-4  781MB/s ± 0%  2511MB/s ± 1%  +221.29%  (p=0.004 n=5+6)
Marshal/string#21-4  842MB/s ± 0%  3059MB/s ± 0%  +263.46%  (p=0.004 n=6+5)
```

You'll see a win when the whole string, or the prefix (52246e59bf9ac39111c1eaf8f6d87f37edccdfa3), doesn't need escaping — the larger the prefix, the bigger the speedup. You could refactor it to copy chunks rather than characters (repeatedly call `EscapeIndex` to find the end of the chunk), and have a circuit breaker for strings with lots of escaping (where you'd fallback to a char loop).

I've disabled the optimization for short strings (< 8 chars). The only case where you'll still see a minor hit is cases like [these](https://github.com/segmentio/encoding/blob/00e544d8f30ff569bcc117a868cc046941af8ff4/json/json_test.go#L154-L155).